pyproject.toml
[project]
name = "webdock"
version = "3.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "load-dotenv>=0.1.0",
    "requests>=2.31.0,<3",
    "setuptools>=80.9.0",
]
